pyproject.toml
[project]
name = "hello-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.5.5",
    "dashscope>=1.25.14",
    "deepagents>=0.4.11",
    "docx>=0.2.4",
    "docx2txt>=0.9",
    "fastapi>=0.135.1",
    "ipympl>=0.10.0",
    "jq>=1.11.0",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langchain-tavily>=0.2.17",
    "markdown>=3.10.2",
    "mcp>=1.26.0",
    "milvus>=2.3.9",
    "openai>=2.29.0",
    "pdfminer-six>=20260107",
    "pymilvus>=2.6.10",
    "pymupdf>=1.27.2",
    "pypdf>=6.8.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
    "unstructured[pdf]>=0.21.5",
]

[dependency-groups]
dev = ["ruff>=0.14.7"]
